feat(components/atom/panel): open classNames
