+++ components/atom/panel/src/ColorPanel.js
@@ -2,18 +2,19 @@ import PropTypes from 'prop-types'
 
 import PolymorphicElement from '@s-ui/react-primitive-polymorphic-element'
 
-import {ALPHA, BORDER_RADIUS, COLORS, ELEVATION, getClassNames} from './constants.js'
+import {ALPHA, BORDER_RADIUS, COLORS, DEFAULT_ALPHA, DEFAULT_COLOR, ELEVATION, getColorClassNames} from './settings.js'
 
-const ColorPanel = function ({
+const ColorPanel = ({
   as = 'div',
-  alpha = ALPHA.CONTRAST,
-  color = COLORS.DEFAULT,
+  alpha = DEFAULT_ALPHA,
+  color = DEFAULT_COLOR,
   children,
   id,
+  className,
   ...otherProps
-}) {
+}) => {
   return (
-    <PolymorphicElement as={as} id={id} className={getClassNames({alpha, color, ...otherProps})}>
+    <PolymorphicElement as={as} id={id} className={getColorClassNames({className, alpha, color, ...otherProps})}>
       {children}
     </PolymorphicElement>
   )
@@ -30,7 +31,8 @@ ColorPanel.propTypes = {
   color: PropTypes.oneOf(Object.values(COLORS)),
   alpha: PropTypes.oneOf(Object.values(ALPHA)),
   rounded: PropTypes.oneOf(Object.values(BORDER_RADIUS)),
-  elevation: PropTypes.oneOf(Object.values(ELEVATION))
+  elevation: PropTypes.oneOf(Object.values(ELEVATION)),
+  className: PropTypes.string
 }
 
 export default ColorPanel
